Create a dedicated spec for array condition rendering
